Give editor its own mouse and keyboard controller.
This will allow us to put more editor-specific functionality
in the editor library (in the controller specifically). For
example, the editor controller has the ability to lock and unlock
the mouse, and automatically tracks first-person facing when the
mouse position is locked, and automatically tracks mouse pointer
position when it's unlocked.

It also wraps keyboard presses and mouse button presses in a nice
way.

Finally, this CL makes the editor use the camera_interface from
component_library.

Change-Id: Iafb6acf286df64cda13fe04b73a4c387f74a92cc
Tested: Ran editor in Zooshi on Linux.