Editor GUI allows you to edit entity contents.
This GUI uses the Flatbuffers reflection system and FlatUI for
rendering. When you have an entity selected, it shows a sidebar with
the entity definitions for that entity, allowing you to edit them and
commit or revert your changes.

Currently working: Editing scalar values, structs, and strings.
Currently not yet working: Resizing vectors.

Tested: Working on Mac and Linux.

Change-Id: I590039ee5af6561f3e950bc72f06daba57520c3f