Added the ability to resize vectors in the world editor.
This change necessitates making the concept of an component being
"modified" more complicated; now, the "fields" of a component (the
GUI) can be modified, or the "buffer" of an entity (the FlatBuffer)
can be modified, in separate steps. This allows us to modify the
buffer to add additional array elements without sending those new
(and likely default-value and broken) elements to the entity until
you commit again.

Tested: On Linux. You can resize arrays of scalars and strings; other
types will require support from the Flatbuffer Reflection system.

Change-Id: I37854fd0b7e2f46edf6cb9324331c48ea64363fa