Tweaked color-coding to be more descriptive, added error color.
Error color appears when you modify something to an inappropriate
value (for now, this means badly parsing structs).

When you have modified a field and have committed it to the FlatBuffer
but it has not yet been used by the application, it will be highlighted
a different color to indicate to the user that it hasn't been fully
used.

Client code should call ClearFlatbufferModifiedFlag() when it is fully
using the modified Flatbuffer, which will reset all these fields.

Change-Id: If92973c6276da9826e743a66f22a53c58953bfdd