Added option for quake-style movement.
The default value is set in the config, but you can change it in the
Settings tab while Scene Lab is enabled.

The option is called Ground Parallel Camera since I can't think of a
better name for it.

If you enable this mode, then using W and S will move you in the camera
direction (including any up or down) rather than keeping you at the same
distance from the ground.

Tested: On Linux desktop.

Change-Id: I2f8ba7e0526b24d725aa6c0f9ee6039661806927