Sample Scene Lab project. A few random cubes in the world.
Sample now builds and runs on Linux and Mac desktop. Android support (for
the game part only, not the editor) coming in a future checkin.

This CL also includes a few small Scene Lab fixes that became necessary
when adding the sample: waking up physics on exiting the editor, and
removing an unnecessary const from returning the camera interface as you
are indeed allowed to modify the camera you get from GetCamera.

It also adds a basic camera that Scene Lab will use by default, if you
don't give it your own camera.

This also includes a change in the way that FlatBuffers build rules are
generated, which depends on I181684b9 in the FlatBuffers library.

Change-Id: I566b884da49f4a458db7601eb1bc2a49b1dd7c90
Tested: Linux desktop. Run `cmake -G"Unix Makefiles" && make` from the
root of scene_lab. Additional changes also tested on Android to ensure
the build still works.