Add the ability to set binary entity file extension.
The config file now has an option to specify the binary file extension
for entity files. By default this is "bin" but your project may have
different file extensions for generated files, so simply set the
binary_entity_file_ext field in your scene lab config file if this is
the case.

Change-Id: I00f4da8738ac99ef87e1e5478bf31fbfee8536d4
Tested: Linux Desktop with Zooshi and sample project.