Added option for adding fields, and version tracking string.
The version tracking string, SceneLab::kVersion, is used by Google
developers to identify which applications uploaded to Google Play
are using this library. See scene_lab.h for more information.

The option to allow adding fields to Flatbuffers is required because
the "add fields" functionality is both experimental and not fully
implemented. It is set to false by default, and users of the Scene
Lab library can turn it on at their own risk.

Change-Id: I53768d79c874a48d68e58f42f6a62cbc383e35bb
Tested: Linux desktop, Scene Lab mode.
Bug: 25742807
Bug: 25742122