Fixing up the Scene Lab docs.
Added `doxygen_layout.xml` to clean up the navigation bar,
fixing a bunch of broken links, added a lot to `README.md` so
that it is more helpful on GitHub, and added `File Reference` pages
for the `File List` via `@file` to make the `API Reference` a bit
nicer.

Tested: Tested on Ubuntu using generate_docs.py and viewed
on Chrome.

Change-Id: Ic24f1fd9aff6426aef7e8757a220c11b6f257daf