Display/parse structs in field order instead of alphabetical.
The reflection Flatbuffer "fields" is actually in alphabetical order,
for quick lookup of field names, but the offset within specifies the
actual struct field order, so we sort the fields by that first.

Previously, many structs (such as a ColorRGBA) would appear in Scene
Lab in a confusing order (ABGR instead of RGB), though notably this
did not affect Vec3 structs (XYZ), which are already sorted.

Bug: 26878009
Tested: Tested on Linux, Zooshi. Tested viewing and editing struct
data for a light, specified as ColorRGBA.

Change-Id: I23cbf653dd8b1056428dcec181a7775980f200c5